refactor: use execa, move post install test to system
