+++ package.json
@@ -42,23 +42,20 @@
     "retry-axios": "0.3.2"
   },
   "devDependencies": {
-    "@types/cross-spawn": "^6.0.0",
+    "@types/execa": "^0.9.0",
     "@types/mocha": "^5.2.4",
     "@types/ncp": "^2.0.1",
     "@types/nock": "^9.1.3",
     "@types/node": "^10.1.0",
     "@types/node-fetch": "^2.1.4",
-    "@types/pify": "^3.0.2",
     "@types/tmp": "0.0.33",
     "@types/uuid": "^3.4.4",
     "assert-rejects": "^1.0.0",
-    "await-spawn": "^2.1.1",
     "codecov": "^3.0.2",
-    "cross-spawn": "^6.0.5",
+    "execa": "^1.0.0",
     "gcx": "0.0.3",
     "googleapis": "^35.0.0",
     "gts": "^0.9.0",
-    "hard-rejection": "^1.0.0",
     "mocha": "^5.2.0",
     "ncp": "^2.0.0",
     "nock": "^10.0.0",
