+++ system-test/fixtures/kitchen/src/index.ts
