+++ test/mocha.opts
@@ -1,4 +1,3 @@
---require hard-rejection/register
 --require source-map-support/register
 --timeout 10000
 --throw-deprecation
